Add evidence<handlerWorker class, and start 3 workers for parallel processing of evidence when slips starts
